Update types to remove signed-unsigned warning during build.

+++ src/main/cpp/loader/omicsds_loader.cc
@@ -29,7 +29,7 @@
 
 std::vector<std::string> split(std::string str, std::string sep) {
   std::vector<std::string> retval;
-  int index;
+  size_t index;
 
   if(str.length() >= 2) {
     if(str[0] == '[') {
@@ -50,7 +50,7 @@ bool FileUtility::generalized_getline(std::string& retval) {
   retval = "";
 
   while(chars_read < file_size || str_buffer.size()) {
-    int idx = str_buffer.find('\n');
+    size_t idx = str_buffer.find('\n');
     if(idx != std::string::npos) {
       retval = retval + str_buffer.substr(0, idx); // exclude newline
       str_buffer.erase(0, idx + 1); // erase newline
@@ -125,7 +125,7 @@ void read_sam_file(std::string filename) {
 
     char *qseq = (char *)malloc(len);
     
-    for(int i=0; i<len; i++){
+    for(size_t i=0; i<len; i++){
       qseq[i] = seq_nt16_str[bam_seqi(q,i)]; //gets nucleotide id and converts them into IUPAC id.
     }
     
@@ -423,7 +423,7 @@ std::vector<OmicsCell> SamReader::get_next_cells() {
     int32_t tlen = m_align->core.isize;
     std::vector<char> qseq(len);
 
-    for(int i=0; i< len ; i++){
+    for(size_t i=0; i< len ; i++){
       qseq[i] = seq_nt16_str[bam_seqi(q,i)]; //gets nucleotide id and converts them into IUPAC id.
     }
 
@@ -589,7 +589,7 @@ bool MatrixReader::parse_next(std::string& sample, std::string& gene, float& sco
     m_current_token = toks[0];
 
     try {
-      for(int i = 0; i < m_columns.size(); i++) {
+      for(size_t i = 0; i < m_columns.size(); i++) {
         m_row_scores[i] = std::stof(toks[i + 1]);
       }
     }
